Extracted ComponentFactory
