+++ src/main/java/views/gui/components/panels/TicketPanel.java
@@ -7,6 +7,7 @@ import database.Property;
 import models.Person;
 import models.Ticket;
 import models.TicketCategory;
+import views.gui.components.ComponentFactory;
 import views.gui.styles.Style;
 
 import javax.swing.*;
@@ -405,7 +406,7 @@ public class TicketPanel extends JPanel implements ListSelectionListener, Proper
         for (Long key : distribution.keySet()) {
             Optional<Person> optDistributionHolder = personDatabase.getById(key);
             if (optDistributionHolder.isEmpty()) continue;
-            Box row = componentFactory.getSmallRow(optDistributionHolder.get(), distribution.get(key));
+            Box row = getSmallRow(optDistributionHolder.get(), distribution.get(key));
             distributionContainer.add(row);
         }
         return distributionContainer;
@@ -545,14 +546,14 @@ public class TicketPanel extends JPanel implements ListSelectionListener, Proper
         List<Double> values = inputs.values().stream().map(field -> Double.parseDouble(field.getText())).toList();
         double sumOfValues = values.stream().reduce(0., Double::sum);
         double diff = abs(sumOfValues - ticket.getCost());
-        JOptionPane.showMessageDialog(null, "Total amount: %.2f. Need %.2f more".formatted(sumOfValues,diff));
+        JOptionPane.showMessageDialog(null, "Total amount: %.2f. Need %.2f more".formatted(sumOfValues, diff));
 
         while (diff > 0.01) {
             JOptionPane.showConfirmDialog(null, message, "Change distribution: total cost %.2f".formatted(costRounded), JOptionPane.DEFAULT_OPTION);
             values = inputs.values().stream().map(field -> Double.parseDouble(field.getText())).toList();
             sumOfValues = values.stream().reduce(0., Double::sum);
             diff = abs(sumOfValues - ticket.getCost());
-            JOptionPane.showMessageDialog(null, "Total amount: %.2f. Need %.2f more".formatted(sumOfValues,diff));
+            JOptionPane.showMessageDialog(null, "Total amount: %.2f. Need %.2f more".formatted(sumOfValues, diff));
         }
 
         for (Long key : keys) {
@@ -660,6 +661,30 @@ public class TicketPanel extends JPanel implements ListSelectionListener, Proper
         return decimalTextField;
     }
 
+    public Box getSmallRow(Person distributionHolder, double amount) {
+        Box box = Box.createHorizontalBox();
+        box.setOpaque(true);
+        box.setBackground(this.style.getTransparantColor());
+        box.setMaximumSize(new Dimension(screenSize.width / 2, 50));
+        box.setAlignmentX(Component.LEFT_ALIGNMENT);
+        box.setAlignmentY(Component.CENTER_ALIGNMENT);
+        box.add(Box.createHorizontalStrut(3 * horizontalOffset));
+
+        JLabel ticketLabel = componentFactory.getSecondarySmallLabel(distributionHolder.getName() + ":");
+        ticketLabel.setMaximumSize(new Dimension(300, 50));
+
+        double amount_rounded = round(amount * 100) / 100.;
+        JLabel distributionLabel = componentFactory.getSecondarySmallLabel("€ " + amount_rounded);
+        distributionLabel.setMaximumSize(new Dimension(200, 50));
+
+        box.add(ticketLabel);
+        box.add(distributionLabel);
+
+        distributionLabel.setForeground(Color.red);
+
+        return box;
+    }
+
     // Private object so that we can pass a Ticket in to the list, but only show its name and id
     private static class TicketListCellRenderer extends DefaultListCellRenderer {
         @Override
@@ -674,113 +699,4 @@ public class TicketPanel extends JPanel implements ListSelectionListener, Proper
             return c;
         }
     }
-
-    private class ComponentFactory {
-        Style style;
-
-        public ComponentFactory(Style style) {
-            this.style = style;
-        }
-
-        public Box getSmallRow(Person distributionHolder, double amount) {
-            Box box = Box.createHorizontalBox();
-            box.setOpaque(true);
-            box.setBackground(this.style.getTransparantColor());
-            box.setMaximumSize(new Dimension(screenSize.width / 2, 50));
-            box.setAlignmentX(Component.LEFT_ALIGNMENT);
-            box.setAlignmentY(Component.CENTER_ALIGNMENT);
-            box.add(Box.createHorizontalStrut(3 * horizontalOffset));
-
-            JLabel ticketLabel = getSecondarySmallLabel(distributionHolder.getName() + ":");
-            ticketLabel.setMaximumSize(new Dimension(300, 50));
-
-            double amount_rounded = round(amount * 100) / 100.;
-            JLabel distributionLabel = getSecondarySmallLabel("€ " + amount_rounded);
-            distributionLabel.setMaximumSize(new Dimension(200, 50));
-
-            box.add(ticketLabel);
-            box.add(distributionLabel);
-
-            distributionLabel.setForeground(Color.red);
-
-            return box;
-        }
-
-        public JLabel getPrimaryNormalLabel(String text) {
-            JLabel label = new JLabel(text);
-            label.setForeground(this.style.getLabel2ForegroundColor());
-            label.setBackground(this.style.getTransparantColor());
-            label.setFont(this.style.getTextFont());
-            label.setHorizontalAlignment(SwingConstants.LEFT);
-            label.setAlignmentX(Component.LEFT_ALIGNMENT);
-            label.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return label;
-        }
-
-        public JLabel getSecondaryNormalLabel(String text) {
-            JLabel label = new JLabel(text);
-            label.setForeground(this.style.getLabel1ForegroundColor());
-            label.setBackground(this.style.getTransparantColor());
-            label.setFont(this.style.getTextFont());
-            label.setHorizontalAlignment(SwingConstants.LEFT);
-            label.setAlignmentX(Component.LEFT_ALIGNMENT);
-            label.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return label;
-        }
-
-        public JLabel getPrimarySmallLabel(String text) {
-            JLabel label = new JLabel(text);
-            label.setForeground(this.style.getLabel2ForegroundColor());
-            label.setBackground(this.style.getTransparantColor());
-            label.setFont(this.style.getSmallTextFont());
-            label.setHorizontalAlignment(SwingConstants.LEFT);
-            label.setAlignmentX(Component.LEFT_ALIGNMENT);
-            label.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return label;
-        }
-
-        public JLabel getSubtitleLabel(String text) {
-            JLabel label = new JLabel(text);
-            label.setForeground(this.style.getLabel1ForegroundColor());
-            label.setBackground(this.style.getTransparantColor());
-            label.setFont(this.style.getBoldSubtitleFont());
-            label.setHorizontalAlignment(SwingConstants.LEFT);
-            label.setAlignmentX(Component.LEFT_ALIGNMENT);
-            label.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return label;
-        }
-
-        public JLabel getSecondarySmallLabel(String text) {
-            JLabel label = new JLabel(text);
-            label.setForeground(this.style.getLabel1ForegroundColor());
-            label.setBackground(this.style.getTransparantColor());
-            label.setFont(this.style.getSmallTextFont());
-            label.setHorizontalAlignment(SwingConstants.LEFT);
-            label.setAlignmentX(Component.LEFT_ALIGNMENT);
-            label.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return label;
-        }
-
-        public JButton getPrimaryButton(String text) {
-            JButton button = new JButton(text);
-            button.setBackground(style.getButton1ForegroundColor());
-            button.setForeground(style.getButton1BackgroundColor());
-            button.setFont(style.getButtonFont());
-            button.setAlignmentX(Component.LEFT_ALIGNMENT);
-            button.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return button;
-        }
-
-        public JButton getSecondaryButton(String text) {
-            JButton button = new JButton(text);
-            button.setBackground(style.getButton2ForegroundColor());
-            button.setForeground(style.getButton2BackgroundColor());
-            button.setFont(style.getButtonFont());
-            button.setAlignmentX(Component.LEFT_ALIGNMENT);
-            button.setAlignmentY(Component.CENTER_ALIGNMENT);
-            return button;
-        }
-    }
-
-
 }
